requirements.txt
annotated-types==0.7.0
anyio==4.10.0
backtrader==1.9.78.123
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
distro==1.9.0
Flask==3.1.2
fonttools==4.59.2
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langsmith==0.4.25
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.6
multitasking==0.0.12
numpy==2.3.2
openai==1.106.1
orjson==3.11.3
packaging==25.0
pandas==2.3.2
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
protobuf==6.32.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.65
zstandard==0.24.0
